// src/components/FormInput/form.d.ts
export interface formProps {
    defaultValue: string,
    group: string,
    label: string,
    patatern: string,
    required: true,
    type: string,
    variable: string,
    description: string,
    pattern: string,
    message: string,
    cluster: any,
    namespace: any,
    default: string,
    field: any
}